requirements.txt
certifi==2019.9.11
chardet==3.0.4
codecov==2.0.15
coverage==4.5.4
dj-database-url==0.5.0
Django==2.0.13
django-heroku==0.3.1
gunicorn==20.0.0
idna==2.8
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
whitenoise==4.1.4
djangorestframework==3.10.3
django-ratelimit
Faker